Rename ScalarType --> BaseScalarType
