+++ src/ts/shared/schemaType/BaseDateType.ts
@@ -1,7 +1,7 @@
-import ScalarType from './ScalarType';
+import BaseScalarType from './BaseScalarType';
 
 /** Base Type for all date types */
-abstract class BaseDateType extends ScalarType {
+abstract class BaseDateType extends BaseScalarType {
 	/** Transform query constants to internal u2 date format */
 	public override transformToQuery(value: unknown): unknown {
 		return value === '' || value == null ? '' : this.transformToDb(value);
